+++ eq-author-api/schema/resolvers/utils/index.js
@@ -1,3 +1,4 @@
 const getters = require("./getters");
+const helpers = require("./helpers");
 
-module.exports = { ...getters };
+module.exports = { ...getters, ...helpers };
